Refactor: auth 관련 컴포넌트 수정
